Move parsetable.ParseTableProduction.getAllProductionLabels to grammar.AssociativityInfo
